Quiz 4 - practical machine learning

+++ practicalmachinelearning/Project/Project.R
@@ -13,16 +13,23 @@
 # 2 You should also apply your machine learning algorithm to the 20 test cases available
 #     in the test data above. Please submit your predictions in appropriate format to the
 #     programming assignment for automated grading. See the programming assignment for additional details.
-require(dplyr);
 
+
+require(dplyr);library(data.table);
+library(RCurl)
+require(caret);require(randomForest);require(e1071);require(ada);
 require(rattle);require(rpart.plot);
 
+if(!file.exists("train.rds")){
+  trainURL<-getURL("https://d396qusza40orc.cloudfront.net/predmachlearn/pml-training.csv")
   trainAll<-read.csv(text=trainURL)
   saveRDS(trainAll,"train.rds")
 } else {
   trainAll<-readRDS("train.rds")
 }
 
+if(!file.exists("test.rds")){
+  testURL<-getURL("https://d396qusza40orc.cloudfront.net/predmachlearn/pml-testing.csv")
   testingFinal<-read.csv(text=testURL)
   saveRDS(testingFinal,"test.rds")
 } else {
@@ -39,40 +46,41 @@ tidyTrain$user_name<-NULL
 tidyTrain$num_window<-NULL
 # remove timestamp columns from dataset
 tidyTrain<-tidyTrain[,-grep("time",names(tidyTrain))]
-testing <- read.csv("testing.csv")
+
 ## -----------Choosing predictors----------
 ## Which columns have minimal variance
 # Do analysis using nearZeroVar function based on caret preprocessing tutorial at  http://topepo.github.io/caret/preprocess.html
+nzv<-nearZeroVar(tidyTrain)
 names(tidyTrain)[nzv]
 # #This shows that new_window has near zero variance but i feel this is important as it is the start of a rep
 # colsToEliminate=nzv[-1]
 tidyTrain<-(tidyTrain[,-nzv])
-# Set a seed so results are reproducible
+
 ## Remove columns that are mostly NAs - more than 97% NA
 apply(tidyTrain,2,function(x){sum(is.na(x))/length(x)})
 nullMoreThan97<-apply(tidyTrain,2,function(x){as.logical(sum(is.na(x))/length(x) > 0.975)})
 tidyTrain<-tidyTrain[,-which(nullMoreThan97)]
-set.seed(32323)
-folds <- createFolds(y=data$classe,k=10,
+
+
 ## Partition training set further into training and test
 trainIndex <- createDataPartition(y=tidyTrain$classe, p=0.75, list=FALSE)
 training<- tidyTrain[trainIndex,]
 testing<- tidyTrain[-trainIndex,]
-                     list=TRUE,returnTrain=TRUE)
+
 ## Fit a Decision Tree model to the data
 rpMdl<-train(classe~.,method="rpart",data=training)
 #fancyRpartPlot(rpMdl$finalModel)
 testing$predicted <- predict(rpMdl, newdata = testing)
 confusionMatrix(testing$predicted,testing$classe)
 #Shows accuracy of ~52% - terrible
-#Check if any columns can be eliminated because they are all NA
+
 ## Fit a Decision Tree rpart2 model to the data
 rpMdl<-train(classe~.,method="rpart2",data=training)
 #fancyRpartPlot(rpMdl2$finalModel)
 testing$predicted <- predict(rpMdl, newdata = testing)
 confusionMatrix(testing$predicted,testing$classe)
 #Shows a better accuracy of 65% better
-which(sapply(data, function(x){all(is.na(x))}))
+
 #Fit a randomForest model to data
 rfMdl<-train(classe~.,method="rf",data=training)
 testing$predicted<-predict(rfMdl, newdata = testing)
@@ -80,8 +88,11 @@ table(testing$predicted,testing$classe)
 confusionMatrix(testing$predicted,testing$classe)
 #Random forest model had an accuracy of 99%
 
-//Lets try k-fold (10 fold) cross validation
+
+#fitControl <- trainControl(method = "repeatedcv",number = 10,  repeats = 10)
 
 #For the final model - lets train it on the entire testing dataset
-    preProcess = c("center","scale"),
+  rfMdlFinal <-train(classe~., method="rf",data=tidyTrain)
+
+# Predict testing set with Random Forest Model
 testingFinal$classe<-predict(rfMdlFinal,newdata=testingFinal)
